Move a pasta Exceptions para o projeto 'Propeus.Modulo.Abstrato'
