+++ src/Propeus.Modulo.Abstrato/Exceptions/ModuloRegistradoException.cs
@@ -1,18 +1,18 @@
 ﻿using System;
 using System.Runtime.Serialization;
 
-namespace Propeus.Modulo.Core.Exceptions
+namespace Propeus.Modulo.Abstrato.Exceptions
 {
     /// <summary>
-    /// Excecao para quando o modulo informado nao foi encontrado
+    /// Excecao para quando o mesmo modulo for registrado mais de uma unica vez
     /// </summary>
     [Serializable]
-    public class ModuloNaoEncontradoException : Exception
+    public class ModuloRegistradoException : Exception
     {
         /// <summary>
         /// Construtor padrao
         /// </summary>
         /// <param name="message">Mensagem do erro</param>
-        public ModuloNaoEncontradoException(string message) : base(message) { }
+        public ModuloRegistradoException(string message) : base(message) { }
     }
 }
