+++ src/Propeus.Modulo.Abstrato/Exceptions/TipoModuloAmbiguoException.cs
@@ -1,18 +1,18 @@
 ﻿using System;
 using System.Runtime.Serialization;
 
-namespace Propeus.Modulo.Core.Exceptions
+namespace Propeus.Modulo.Abstrato.Exceptions
 {
     /// <summary>
-    /// Excecao para quando o modulo nao seguir os padroes de implementacao
+    /// Excecao para quando for encontrado mais de um modulo de mesmo nome
     /// </summary>
     [Serializable]
-    public class TipoModuloInvalidoException : Exception
+    public class TipoModuloAmbiguoException : Exception
     {
         /// <summary>
         /// Construtor padrao
         /// </summary>
         /// <param name="message">Mensagem do erro</param>
-        public TipoModuloInvalidoException(string message) : base(message) { }
+        public TipoModuloAmbiguoException(string message) : base(message) { }
     }
 }
