+++ src/Propeus.Modulo.Abstrato/Exceptions/TipoModuloInvalidoException.cs
@@ -1,18 +1,18 @@
 ﻿using System;
 using System.Runtime.Serialization;
 
-namespace Propeus.Modulo.Core.Exceptions
+namespace Propeus.Modulo.Abstrato.Exceptions
 {
     /// <summary>
-    /// Excecao para quando o modulo for criado fora do Gerenciador
+    /// Excecao para quando o modulo nao seguir os padroes de implementacao
     /// </summary>
     [Serializable]
-    public class ModuloNaoRegistradoException : Exception
+    public class TipoModuloInvalidoException : Exception
     {
         /// <summary>
         /// Construtor padrao
         /// </summary>
         /// <param name="message">Mensagem do erro</param>
-        public ModuloNaoRegistradoException(string message) : base(message) { }
+        public TipoModuloInvalidoException(string message) : base(message) { }
     }
 }
