+++ src/Propeus.Modulo.Abstrato/Exceptions/TipoModuloNaoEncontradoException.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Runtime.Serialization;
 
-namespace Propeus.Modulo.Core.Exceptions
+namespace Propeus.Modulo.Abstrato.Exceptions
 {
     /// <summary>
     /// Excecao para quando o tipo do modulo informado nao for encontrado no Assembly
