+++ src/Propeus.Modulo.Core/Gerenciador.cs
@@ -9,9 +9,9 @@ using System.Threading.Tasks;
 
 using Propeus.Modulo.Abstrato;
 using Propeus.Modulo.Abstrato.Atributos;
+using Propeus.Modulo.Abstrato.Exceptions;
 using Propeus.Modulo.Abstrato.Helpers;
 using Propeus.Modulo.Abstrato.Interfaces;
-using Propeus.Modulo.Core.Exceptions;
 using Propeus.Modulo.Util;
 using Propeus.Modulo.Util.Thread;
